refactor: extract tui view
